requirements/local.txt
-r base.txt

django-debug-toolbar==1.8
django-model-utils==3.0.0
django-extensions==1.9.7